pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-api-rate-limiter"
version = "0.1.0"
description = "Smart API Rate Limiter with JWT, Redis, metrics"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.111.1",
    "uvicorn[standard]==0.24.0",
    "redis==5.0.0",
    "fakeredis==2.21.0",
    "bcrypt==4.0.1",
    "PyJWT==2.8.0",
    "pydantic==2.8.0",
    "pydantic-settings==2.8.0",
    "pytest==8.4.2",
    "pytest-asyncio==0.26.0",
    "httpx==0.28.1",
    "prometheus-fastapi-instrumentator==7.1.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
